requirements.txt
airsim==1.8.1
alphashape==1.3.1
descartes==1.1.0
ground==8.2.0
matplotlib==3.5.1
numpy==1.22.3
open3d==0.15.1
PyYAML==6.0.1
scipy==1.7.3
sect==6.1.0
Shapely==1.8.2
tqdm==4.66.1
